requirement.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
idna==2.8
oauthlib==3.0.1
Pillow==5.3.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.6
requests==2.21.0
requests-oauthlib==1.2.0
rest-social-auth==2.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
uWSGI==2.0.17.1
